another int used to store a pointer, which breaks AMD64; thanks once again to Adrian Schroeter from SuSE for spotting this
CVS patchset: 6332
CVS date: 2004/04/05 19:31:35

+++ src/post/goom/surface.c
@@ -1,10 +1,11 @@
 #include "surface.h"
 #include <stdlib.h>
+#include <inttypes.h>
 
 Surface * surface_new (int w, int h) {
   Surface * s = (Surface*)malloc(sizeof(Surface));
   s->realstart = (int*)malloc(w*h*4 + 128);
-  s->buf = (int*)((int)s->realstart + 128 - (((int)s->realstart) % 128));
+  s->buf = (int*)((uintptr_t)s->realstart + 128 - (((uintptr_t)s->realstart) % 128));
   s->size = w*h;
   s->width = w;
   s->height = h;
